requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
requests
sqlparse
python-multipart
groq
pandas
openpyxl
qdrant-client
pymongo
streamlit
requests
pydantic
fastapi
uvicorn
sqlalchemy
pymysql
psycopg2-binary
pandas
openpyxl
xlrd
sqlparse
python-multipart
pydantic
requests
qdrant-client
pymongo
python-dotenv
